Add gas changes

+++ contracts/interfaces/IUniswapV2Router.sol
@@ -9,22 +9,22 @@ interface IUniswapV2Router {
     function addLiquidity(
         address tokenA,
         address tokenB,
-        uint amountADesired,
-        uint amountBDesired,
-        uint amountAMin,
-        uint amountBMin,
+        uint256 amountADesired,
+        uint256 amountBDesired,
+        uint256 amountAMin,
+        uint256 amountBMin,
         address to,
-        uint deadline
-    ) external returns (uint amountA, uint amountB, uint liquidity);
+        uint256 deadline
+    ) external returns (uint256 amountA, uint256 amountB, uint256 liquidity);
 
     function getAmountsOut(
-        uint amountIn,
+        uint256 amountIn,
         address[] calldata path
-    ) external view returns (uint[] memory amounts);
+    ) external view returns (uint256[] memory amounts);
 
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
         uint deadline
